auth-service/requirements.txt
aiofiles==24.1.0
anyio==4.9.0
asgiref==3.8.1
async-property==0.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
deprecation==2.1.0
Django==5.2.1
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
josepy==2.0.0
jwcrypto==1.5.6
mozilla-django-oidc==4.0.1
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.1.0
python-keycloak==5.5.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
pika==1.3.2